Login/Signup/Verify
Finished the login/signup/verify functions to correctly display the
right fragment based on the “role” of the user logging in. Still need
to update UI for Tenant, but can’t get iOS version to run so need to
get with Bryan to get me some screens to develop correctly.