Maintenance View Implemented
Created and implemented a bit more logical in displaying manager versus
tenant, but still have a lot of design to go through. Added the
Maintenance request view to Tenant, and still need to add the “Create
Request” feature to populate the view more. Basic RecyclerView being
used and still learning how to properly theme the layout to make it
look nice.