Maintenance Request - Beta Submission
Had a set back with Menu display that I was able to overcome with
loading menus in each fragment. With that, I revised a bit of code for
menu actions, and wasn’t able to complete the Maintenance Request view
or adding. Should be working on that tomorrow and get a finished
Request system for the tenant and start working on the Manager view and
Maintenance Staff view tomorrow evening.

Pushing development to Master for Beta submission.