Add Request and Image Capture
Request add finished with image capture working. I still need to
implement the “view” fragment for the items in the recycler view and
format some things if time allows, but the request add works and the
image capture does as well. This is tested on SDK 25 - haven’t tested
for lower yet.