Requests and Complaints
Updated the Maintenance system to have full add and view request
function as well as photo storage. Photo only works on physical device.
Still haven’t tested on SDK < 25.

Complaint system implemented using the same basic code of the request
system, but since complaints will be a one message and done method I
implemented a dialog view to show the content. This should update when
the manager acknowledges the complaint.