Manager Complaints View
Completed the implementation of the complaints view and acknowledgement
for the manager side. Complaints menu now shows links for active or
close complaints and will allow the manager to go to the group that
they would like to see. In active, they can send an “acknowledgement”
and reply only once and then it moves to closed and shows closed status
for the tenant. Tenant will only see the status update and reply if
noted, otherwise they see a date when the manager acknowledged it.