Staff Creation - Account Validation
Managers can now create staff accounts, though I have no staff home
created just yet. Also, after going through the staff creation I added
a feature to allow the user to stop at the verification screen and
return at a later time. This allows the process to continue later if
the user (tenant, staff, or manager) needs to return due to an error in
the application and come back to the verification screen. eg. the phone
was entered wrong or they type in the wrong “verification” code.